+++ src/app/api/send/route.ts
@@ -1,16 +1,10 @@
 import { EmailTemplate } from "@/components/EmailTemplate";
-import { NextResponse } from "next/server";
+import ContactFormData from "@/interface/contactFormData";
+import { NextRequest, NextResponse } from 'next/server';
 import { Resend } from "resend";
 
 const resend = new Resend("re_16F3jeYk_APWSuQEV95DHtPTagbVGkBFn");
 //const resend = new Resend(process.env.RESEND_API_KEY);
-
-interface ContactFormData {
-    firstName: string;
-    number: string;
-    mail: string;
-    message: string;
-  }
 
 // export async function POST() {
 //   try {
@@ -28,17 +22,25 @@ interface ContactFormData {
 //   }
 // }
 
-export async function POST({firstName, number, mail, message}:ContactFormData ) {
+export async function POST(req: NextRequest) {
     try {
-      const responseData = await resend.emails.send({
-        from: "Acme <onboarding@resend.dev>",
-        to: ["desarrollo.happiier@gmail.com"],
-        subject: "Mensaje de contacto",
-        react: EmailTemplate({ firstName, number, mail, message }),
-        text: "",
-      });
-  
-      return NextResponse.json(responseData);
+      // Verifica si req.body es del tipo ContactFormData
+      if (req.body instanceof Object) {
+        const { firstName, number, mail, message } = req.body as unknown as ContactFormData;
+
+        const responseData = await resend.emails.send({
+          from: "Acme <onboarding@resend.dev>",
+          to: ["desarrollo.happiier@gmail.com"],
+          subject: "Mensaje de contacto",
+          react: EmailTemplate({ firstName, number, mail, message }),
+          text: "",
+        });
+
+        return NextResponse.json(responseData);
+      } else {
+        // Manejar el caso donde req.body no es del tipo esperado
+        return NextResponse.json({ error: "El cuerpo de la solicitud no es válido" });
+      }
     } catch (error) {
       return NextResponse.json({ error });
     }
