+++ src/interface/contactFormData.ts
@@ -1,0 +1,8 @@
+interface ContactFormData {
+    firstName: string;
+    number: string;
+    mail: string;
+    message: string;
+}
+
+export default ContactFormData;
